requirements.txt
streamlit
google-generativeai
pandas
numpy
matplotlib
seaborn
python-pptx
chromadb
langchain_community
pypdf